pyproject.toml
[project]
name = "typer-group-prefix"
version = "0.1.5.5"
description = "Prefix handler for typer group options"
readme = "README.md"
authors = [
    { name = "Kalle Møller Aagaard (KM)", email = "git@k-moeller.dk" }
]
requires-python = ">=3.11"
dependencies = [
    "typer-di>=0.1.3",
    "typer-slim>=0.17.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
example = [
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
    "typer-di>=0.1.3",
]
